+++ src/main/kotlin/dev/usbharu/hideout/activitypub/service/common/ActivityPubProcessor.kt
@@ -1,7 +1,6 @@
-package dev.usbharu.hideout.activitypub.service.tmp
+package dev.usbharu.hideout.activitypub.service.common
 
 import dev.usbharu.hideout.activitypub.domain.model.objects.Object
-import dev.usbharu.hideout.activitypub.service.common.ActivityType
 
 interface ActivityPubProcessor<T : Object> {
     suspend fun process(activity: ActivityPubProcessContext<T>)
